+++ resources/js/search-modal.js
@@ -3,15 +3,15 @@ import InstantSearch from 'vue-instantsearch/vue3/es'
 import { container } from 'jenesius-vue-modal'
 
 // Components
-import Meilisearch from './components/organisms/Meilisearch.vue'
+import Meilisearch from './components/InstantsearchProvider.vue'
 import ModalTrigger from './components/organisms/ModalTrigger.vue'
 
 const app = createApp({
   components: {
     Meilisearch,
     VueModalContainer: container,
-    ModalTrigger
-  }
+    ModalTrigger,
+  },
 })
 
 app.use(InstantSearch)
